+++ lib/chopper/chopper.dart
@@ -1,4 +1,6 @@
+export 'base_error.dart';
 export 'color_converter.dart';
 export 'datetime_converter.dart';
+export 'json_converter_exception.dart';
 export 'json_serializable_converter.dart';
 export 'no_content.dart';