+++ test/chopper/json_serializable_converter_test.dart
@@ -1,6 +1,5 @@
-import 'dart:convert';
-
 import 'package:chopper/chopper.dart';
+import 'package:dcc_toolkit/chopper/json_converter_exception.dart';
 import 'package:dcc_toolkit/chopper/json_serializable_converter.dart';
 import 'package:dcc_toolkit/chopper/no_content.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -14,7 +13,7 @@ void main() {
 
     const converter = JsonSerializableConverter({});
 
-    expect(() => converter.convertResponse<TestModel, TestModel>(response), throwsA(isA<JsonUnsupportedObjectError>()));
+    expect(() => converter.convertResponse<TestModel, TestModel>(response), throwsA(isA<JsonConverterException>()));
   });
 
   test('Error is thrown when wrong factory is added', () {
@@ -22,7 +21,7 @@ void main() {
 
     const converter = JsonSerializableConverter({TestModel: TestModel2.fromJson});
 
-    expect(() => converter.convertResponse<TestModel, TestModel>(response), throwsA(isA<JsonUnsupportedObjectError>()));
+    expect(() => converter.convertResponse<TestModel, TestModel>(response), throwsA(isA<JsonConverterException>()));
   });
 
   test('Uses model to convert json into model', () async {